pyproject.toml
[tool.poetry]
name = "vm-api"
version = "0.1.0"
description = "VM API with observability and authentication"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "^0.7.0"
anyio = "^4.10.0"
certifi = "^2025.8.3"
charset-normalizer = "^3.4.3"
click = "^8.2.1"
distro = "^1.9.0"
dotenv = "^0.9.9"
exceptiongroup = "^1.3.0"
fastapi = "^0.116.1"
groq = "^0.31.1"
h11 = "^0.16.0"
httpcore = "^1.0.9"
httpx = "^0.28.1"
idna = "^3.10"
psycopg2-binary = "^2.9.0"
protobuf = ">=5.0,<6.0"
pydantic = "^2.11.7"
pydantic_core = "^2.33.2"
email-validator = "^2.1.0"
python-dotenv = "^1.1.1"
pytz = "^2025.2"
requests = "^2.32.5"
slack_sdk = "^3.36.0"
sniffio = "^1.3.1"
starlette = "^0.47.3"
typing-inspection = "^0.4.1"
typing_extensions = "^4.15.0"
urllib3 = "^2.5.0"
uvicorn = "^0.35.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic-settings = "^2.10.1"
boto3 = "^1.35.0"
aioboto3 = "^13.1.0"
langchain = "^0.3.27"
langchain-groq = "^0.3.8"
langchain-openai = "^0.3.0"  # For BYOLLM (ChatOpenAI, AzureChatOpenAI)
langchain-google-genai = "^2.0.8"
langchain-core = "^0.3.78"
langgraph = "^0.2.0"
openai = "^1.0.0"  # For BYOLLM verification (OpenAI and Azure OpenAI)
google-generativeai = "^0.8.0"  # For BYOLLM verification (Gemini)
greenlet = "^3.2.4"
langfuse = "^2.0.0"
redis = "^5.0.0"
slowapi = "^0.1.9"
tenacity = "^9.0.0"
sentry-sdk = {version = ">=2.35.0", extras = ["fastapi"]}
pillow = "^12.0.0"
stripe = "^11.0.0"

# OpenTelemetry - OTLP exporters for metrics and logs
opentelemetry-api = "^1.28.0"
opentelemetry-sdk = "^1.28.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.28.0"
opentelemetry-instrumentation-fastapi = "^0.49b0"
opentelemetry-instrumentation-logging = "^0.49b0"
argon2-cffi = "^25.1.0"
python-multipart = "^0.0.21"
python-magic = "^0.4.27"
pypdf2 = "^3.0.1"
presidio-analyzer = {version = ">=2.2.360,<3.0.0", markers = "python_version < \"3.14\""}
presidio-anonymizer = {version = ">=2.2.360,<3.0.0", markers = "python_version < \"3.14\""}
en-core-web-lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
black = "^24.0.0"
isort = "^5.13.0"
ruff = "^0.6.9"
mypy = "^1.11.2"
bandit = "^1.7.9"
pip-audit = "^2.7.3"
pre-commit = "^3.8.0"
pytest-cov = "^6.0.0"
aiosqlite = "^0.22.1"

[tool.mypy]
explicit_package_bases = true
mypy_path = "."
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"